Refine OSC version bump workflow
Bumping OSC version is a bit tricky as it should happen at many
locations scattered everywhere in the tree. Commit e2a4c34940fb
introduced an `OSC_VERSION` tag to help tracking these locations.

Unfortunately, `make bundle` wreaks havoc in this logic as it
overwrites lines at several locations and drops the tag:
- everywhere in the `bundle` directory
- some locations in the base CSV
- a single location in `config/manager/kustomization.yaml`

As a result, we can only rely on `OSC_VERSION` for a subset of
the locations that need care. The other ones are:
- the `olm.skipRange` annotation in the base CSV
- the `spec.version` field in the base CSV
- the `spec.replaces` field in the base CSV

`make bundle` should then be run to propagate the version bump
to the rest of the repo.

Document all that.

While here set the tag for the metrics-server container that was
recently added.

Signed-off-by: Greg Kurz <groug@kaod.org>