Fix detection of SNO or converged cluster
An SNO or converged cluster doesn't have any nodes
in the `worker` MachineConfigPool. All nodes are part
of the `master` MachineConfigPool. We use this aspect
to detect SNO or converged cluster to install Kata with CoCo
support.

A true multi-node cluster with have at least one node in the
`worker` MachineConfigPool.
For true multi-node cluster, we expect the admin to label the
target worker node for CoCo installation with specific label (eg.
coco-bm=true) and specify this label in the BM_NODE_LABEL env variable
(eg. export BM_NODE_LABEL="coco-bm=true") before starting the
installation

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>