Strenghten ODK/OCP detection heuristic
The desired image is enough to differentiate OKD and OCP when the cluster
is deployed using default public repositories. It doesn't fly when pulling
from custom repositories though, e.g. in some CI environments. For such
cases, an option is to look for `ostree/rhcos` in the kernel command line,
which is always exposed to containers.

Existing heuristic based on desired image is kept since it works for
most deployments and it is recommended by the OKD team. An extra check
targetting CentOS Stream CoreOS (SCOS) is added for completeness.

The environment based hidden knob is also kept in case it is needed
again in the future.

Signed-off-by: Greg Kurz <groug@kaod.org>