must-gather: updates for the Dockerfile
The Dockerfile for the must-gather image is very outdated.
It actually fails to build because it's still based on centos:7,
which repositories are not available anymore.

Since it doesn't seem to be used, I am suggesting these changes,
which gets closer to what we actually need to release in Openshift.

Changes include:
- use ubi9 minimal for the base image
- do not install additional tools - they're not referenced by our gathering
  scripts, and they've never been installed in our currently released
  images.

Signed-off-by: Julien Ropé <jrope@redhat.com>