Check the existence of the PeerPodConfig CRD
PR #533 added some logic to move away from the PeerPodConfig CRD.
Unfortunately it doesn't cope with the case where the CRD doesn't
exist at all in the cluster. This causes the controller to enter
a reconcile failure loop with the following error.

Failed to migrate PeerPodConfig limit	{"err": "no matches for
 kind \"PeerPodConfig\" in version \"confidentialcontainers.org/v1alpha1\""}

If the CRD isn't defined, no need to migrate obviously.

Signed-off-by: Greg Kurz <groug@kaod.org>