podvm: Add a preStop hook to cleanup left over packer VMs
A script to clean up left over packer resources is added and executed
as part of preStop hook in the pod vm generation job.
The packer resources uses well-defined prefix, so it makes it easier
to query the left over resources and delete them.

The logs are redirected to stdout of PID 1 so that it's accessible
via oc logs command

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>