kata installation should not watch the 'worker' MCP state
When the MCO migrates nodes between pools, the target pool enters the
Updating state and has nodes gradually added to it as they are updated and
rebooted while the source pool is drained by simply setting its machine
count to the desired value at once and never leaves the Updated state.

Because the "worker" pool is the source pool in kata installation, it
follows that nothing can be gained by watching its status.

Normally it doesn't hurt either but if "worker" is Updating already for
some unrelated reason when kata installation is starting, it throws off
the isInstallationInProgress detection and thus the whole installation
flow.

Signed-off-by: Pavel Mores <pmores@redhat.com>