podvm: Select subscription after Azure login
Experimentally, if you have multiple Azure subscription and the one
you are working in is not the first one returned by `az login`, the
next step in the image builder script fails creating the gallery,
because it looks for the resource group in the wrong subscription:

```
ERROR: (ResourceGroupNotFound) Resource group 'c3d-konflux-rg' could not be found.
145
Code: ResourceGroupNotFound
```

This is addressed here by explicitly selecting the subcription with
`az account set --subscription ${AZURE_SUBSCRIPTION_ID}`.

Fixes: #606

Signed-off-by: Christophe de Dinechin <dinechin@redhat.com>