podvm: Restrict Azure image version is a 32-bit integer
The Azure image version is of the form Major(int).Minor(int).Patch(int)
format, where int is between 0 and 2,147,483,647 (inclusive). The
previous code converted the cluster ID (which is in hex format) directly
to an integer and used it as the major version number. However the
converted value could be too large resulting in image creation failure.
This change limits the major version to the max signed int value.  As
for the Minor and Patch values, we use a fixed value of 0.1

Fixes: #KATA-3741
Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>