konflux: work around for controller-gen dependency in hermetic builds
When running in a hermetic environment (no access to the network), the build
was failing because of our Makefile trying to install controller-gen at build
time.

As this line was added by the operator-sdk, we would like to avoid modifying
it. It's actually useful in local builds, to make sure we're using the right
tools.
We also want to avoid using a different Makefile or Dockerfile for Konflux.

This commit provides a way to allow using the same process for building
locally and in our CI. It mimics the behavior we have in CPaaS, relying on
having controller-gen defined as a dependency in our go.mod file.

I'm adding a script that runs "go run" for controller-gen, and place it where
the Makefile expects to find the actual controller-gen binary
'make' will find that and skip the install of the tool, then call that script
in place of the actual tool whenever needed.

- in the context of a local build, "go run" will download and build the tool,
  just like it would if you used "go install", then run it as expected.
- in the context of Konflux in a hermetic environment, "go run" will find the
  controller-gen sources pre-fetched by cachi2/Hermeto, and build it from there.

Signed-off-by: Julien Ropé <jrope@redhat.com>