container builds: controller-gen version management
When we update the controller-tools version, we need to update
go.mod, Makefile and Dockerfile, making sure they're all in sync.

This commit tries to automate that for container builds:
the Dockerfile will create an environment variable with the version
found in go.mod, and use that for setting up the controller-gen script,
and calling 'make' with that version.

For local builds, where only the Makefile is used, the default version
found in the Makefile still needs to be updated. But for container builds,
the value in the Makefile will be overriden, and always in sync with go.mod.

Signed-off-by: Julien Ropé <jrope@redhat.com>