konflux: update CSV references to match the latest snapshot
Since we didn't get bundle updates through nudges in konflux in a while, there
is a mismatch between the reference we have in the CSV and the image used in the
latest build (snapshot) of Konflux.

Because of that, the image is not reachable and we get a failure for the release
process.

Rather that waiting for the next push to podvm-payload to get a nudge, I'm fixing
it by copying the SHA that I can see in the latest release snapshot that raised
a failure for podvm-payload.

Signed-off-by: Julien Ropé <jrope@redhat.com>