tests: add mirrors for konflux images
Konflux images are located in quay.io/redhat-user-workloads/ose-osc-tenant/[...]
but our bundle references images from registry.redhat.io.

To be able to test our bundle with unreleased images, we need to set some mirrors
in the test cluster to check the images from quay instead of registry.redhat.io.

This yaml provides the required mirrors, and needs to be applied to the test cluster.

Signed-off-by: Julien Ropé <jrope@redhat.com>