aws: automate prebuilt artifact upload
Previously, uploading a disk image to an AWS AMI required two manual steps:
1. Creating an S3 bucket
2. Creating the vmimport role/service

This commit leverages the existing ami-helper.sh script to perform these
operations automatically. The functionality is now integrated into the
`aws-podvm-image-handler.sh` script to enable full automation.

This change requires extended permissions for the credentials used,
and is intended to be better integrated in the future.