test-fbc: run the render.sh script as part of the build.
We can "nudge" the test-fbc component when the bundle is rebuilt.
But that will just update the bundle reference, not all the related images.

Rather than nuding this component from all images, I'm running the "render.sh"
script when we build this FBC. This way, when the bundle reference is updated,
the catalog should be updated and built with all related images.

The render.sh script runs "opm" to get the signature of all related images.
Since we can't do that in a hermetic environment, I am disabling hermetic build
as part of this commit too.
This is not an issue for the test-fbc catalog, as it is not meant for release
anyway, only for internal testing.

Signed-off-by: Julien Ropé <jrope@redhat.com>